mfd: zl3073x: Add support to write HW register sequences
Add functionality to write sequence of masked values into specific
HW registers with optional delay after each write.

Each sequence item has an address of HW register to update, value to be
written, mask that specifies which bits should be updated and wait
value that contains number of miliseconds to wait after the register
update.

This functionality is used in subsequent patch that loads and executes
flash utility.

Signed-off-by: Ivan Vecera <ivecera@redhat.com>