mfd: zl3073x: Implement flashing sectors and use it for firmware1
The image type firmware1 uses method of flashing sectors. This method
is used for images that are too large to flash them at once because
there is a size limit for each flash write cycle. Larger images need
to be flashed in parts.

So implement this most complicated method and use it for firmware1
image type.

Signed-off-by: Ivan Vecera <ivecera@redhat.com>