dpll: zl3073x: Add support to get phase offset on input pins
Add support to get phase offset for the input pins. Implement the appropriate
callback that performs DPLL to reference phase error measurement and reports
the measured value. If the DPLL is currently locked to different reference
with higher frequency then the phase offset is modded to the period of
the signal the DPLL is locked to.

Co-developed-by: Prathosh Satish <Prathosh.Satish@microchip.com>
Signed-off-by: Prathosh Satish <Prathosh.Satish@microchip.com>
Signed-off-by: Ivan Vecera <ivecera@redhat.com>