dpll: zl3073x: Read pin supported frequencies from firmware
Load the frequencies from the property provided by the firmware (DT, ACPI...)
that specify what frequencies are supported / allowed for the particular
pins with respect to hardware wiring.
Do not blindly trust firmware and filter out frequencies that cannot be
configured represented in device (input frequencies have to be
factorized by one of the base frequencies and output frequencies have to
divide configured synthesizer frequency).

Signed-off-by: Ivan Vecera <ivecera@redhat.com>